requirements.txt
Django==5.2.5
pandas==2.0.3
openpyxl==3.1.2
django-bootstrap5==23.3
django-extensions==3.2.3
djangorestframework==3.14.0
django-filter==23.2
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0